Move into its own module, apply fixes
